Moved useZAuctionHook logic inside BidProvider
